changes some int to size_t - code compiles with no warnings!!!

+++ hw3/Battleship3D/AlgorithmLoader.cpp
@@ -45,7 +45,7 @@ void AlgorithmLoader::exportAlgos(std::vector<GetAlgoFuncType>& algos)
 	}
 }
 
-int AlgorithmLoader::size() const
+size_t AlgorithmLoader::size() const
 {
 	return libs.size();
 }
